LPD-41296 Refactor
